Move things among files and several other changes according to @nrc's advice
