+++ src/storage/txn/commands/check_txn_status.rs
@@ -3,8 +3,9 @@
 use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
 use crate::storage::mvcc::MvccTxn;
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand};
-use crate::storage::txn::process::{ReleasedLocks, WriteResult};
+use crate::storage::txn::commands::{
+    Command, CommandExt, ReleasedLocks, TypedCommand, WriteCommand, WriteResult,
+};
 use crate::storage::txn::Result;
 use crate::storage::{ProcessResult, Snapshot, Statistics, TxnStatus};
 use kvproto::kvrpcpb::ExtraOp;
