+++ src/storage/txn/commands/mod.rs
@@ -40,22 +40,26 @@ pub use txn_heart_beat::TxnHeartBeat;
 #[cfg(test)]
 pub(crate) use prewrite::FORWARD_MIN_MUTATIONS_NUM;
 
+pub use resolve_lock::RESOLVE_LOCK_BATCH_SIZE;
+
 use std::fmt::{self, Debug, Display, Formatter};
 use std::iter::{self, FromIterator};
 use std::marker::PhantomData;
 
 use kvproto::kvrpcpb::*;
-use txn_types::{Key, TimeStamp};
+use txn_types::{Key, TimeStamp, Value, Write};
 
-use crate::storage::kv::Engine;
+use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::{LockManager, WaitTimeout};
+use crate::storage::mvcc::Lock as MvccLock;
+use crate::storage::mvcc::{MvccReader, ReleasedLock};
 use crate::storage::txn::latch::{self, Latches};
-use crate::storage::txn::process::WriteResult;
 use crate::storage::txn::{ProcessResult, Result};
 use crate::storage::types::{
     MvccInfo, PessimisticLockRes, PrewriteResult, StorageCallbackType, TxnStatus,
 };
-use crate::storage::{metrics, Result as StorageResult, Snapshot, Statistics};
+use crate::storage::{lock_manager, metrics, Result as StorageResult, Snapshot, Statistics};
+use kvproto::kvrpcpb::Context;
 use pd_client::PdClient;
 use std::sync::Arc;
 use tikv_util::collections::HashMap;
@@ -305,6 +309,77 @@ impl From<MvccGetByStartTsRequest> for TypedCommand<Option<(Key, MvccInfo)>> {
     }
 }
 
+#[derive(Default)]
+struct ReleasedLocks {
+    start_ts: TimeStamp,
+    commit_ts: TimeStamp,
+    hashes: Vec<u64>,
+    pessimistic: bool,
+}
+
+pub(super) struct WriteResult {
+    pub ctx: Context,
+    pub to_be_write: WriteData,
+    pub rows: usize,
+    pub pr: ProcessResult,
+    // (lock, is_first_lock, wait_timeout)
+    pub lock_info: Option<(lock_manager::Lock, bool, Option<WaitTimeout>)>,
+}
+
+impl ReleasedLocks {
+    pub fn new(start_ts: TimeStamp, commit_ts: TimeStamp) -> Self {
+        Self {
+            start_ts,
+            commit_ts,
+            ..Default::default()
+        }
+    }
+
+    pub fn push(&mut self, lock: Option<ReleasedLock>) {
+        if let Some(lock) = lock {
+            self.hashes.push(lock.hash);
+            if !self.pessimistic {
+                self.pessimistic = lock.pessimistic;
+            }
+        }
+    }
+
+    // Wake up pessimistic transactions that waiting for these locks.
+    pub fn wake_up<L: LockManager>(self, lock_mgr: &L) {
+        lock_mgr.wake_up(self.start_ts, self.hashes, self.commit_ts, self.pessimistic);
+    }
+}
+
+type LockWritesVals = (
+    Option<MvccLock>,
+    Vec<(TimeStamp, Write)>,
+    Vec<(TimeStamp, Value)>,
+);
+
+fn find_mvcc_infos_by_key<S: Snapshot>(
+    reader: &mut MvccReader<S>,
+    key: &Key,
+    mut ts: TimeStamp,
+) -> Result<LockWritesVals> {
+    let mut writes = vec![];
+    let mut values = vec![];
+    let lock = reader.load_lock(key)?;
+    loop {
+        let opt = reader.seek_write(key, ts)?;
+        match opt {
+            Some((commit_ts, write)) => {
+                ts = commit_ts.prev();
+                writes.push((commit_ts, write));
+            }
+            None => break,
+        };
+    }
+    for (ts, v) in reader.scan_values_in_default(key)? {
+        values.push((ts, v));
+    }
+    Ok((lock, writes, values))
+}
+
 pub trait CommandExt: Display {
     fn tag(&self) -> metrics::CommandKind;
 
@@ -379,8 +454,7 @@ impl Command {
             Command::MvccByStartTs(t) => t,
         }
     }
-
-    pub fn read_command_mut<E: Engine>(&mut self) -> &mut dyn ReadCommand<E> {
+    pub fn read_command_mut<S: Snapshot>(&mut self) -> &mut dyn ReadCommand<S> {
         match self {
             Command::ScanLock(t) => t,
             Command::ResolveLockReadPhase(t) => t,
@@ -390,7 +464,7 @@ impl Command {
         }
     }
 
-    pub fn write_command_mut<S: Snapshot, L: LockManager, P: PdClient + 'static>(
+    pub(super) fn write_command_mut<S: Snapshot, L: LockManager, P: PdClient + 'static>(
         &mut self,
     ) -> &mut dyn WriteCommand<S, L, P> {
         match self {
@@ -470,15 +544,13 @@ impl Debug for Command {
     }
 }
 
-pub trait ReadCommand<E: Engine>: CommandExt {
-    fn process_read(
-        &mut self,
-        snapshot: E::Snap,
-        statistics: &mut Statistics,
-    ) -> Result<ProcessResult>;
+pub trait ReadCommand<S: Snapshot>: CommandExt {
+    fn process_read(&mut self, snapshot: S, statistics: &mut Statistics) -> Result<ProcessResult>;
 }
 
-pub trait WriteCommand<S: Snapshot, L: LockManager, P: PdClient + 'static>: CommandExt {
+pub(super) trait WriteCommand<S: Snapshot, L: LockManager, P: PdClient + 'static>:
+    CommandExt
+{
     fn process_write(
         &mut self,
         snapshot: S,
