+++ src/storage/txn/commands/mvcc_by_start_ts.rs
@@ -1,12 +1,12 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::kv::Engine;
 use crate::storage::mvcc::MvccReader;
-use crate::storage::txn::commands::{Command, CommandExt, ReadCommand, TypedCommand};
-use crate::storage::txn::process::find_mvcc_infos_by_key;
+use crate::storage::txn::commands::{
+    find_mvcc_infos_by_key, Command, CommandExt, ReadCommand, TypedCommand,
+};
 use crate::storage::txn::{ProcessResult, Result};
 use crate::storage::types::MvccInfo;
-use crate::storage::{ScanMode, Statistics};
+use crate::storage::{ScanMode, Snapshot, Statistics};
 use txn_types::{Key, TimeStamp};
 
 command! {
@@ -32,12 +32,8 @@ impl CommandExt for MvccByStartTs {
     gen_lock!(empty);
 }
 
-impl<E: Engine> ReadCommand<E> for MvccByStartTs {
-    fn process_read(
-        &mut self,
-        snapshot: E::Snap,
-        statistics: &mut Statistics,
-    ) -> Result<ProcessResult> {
+impl<S: Snapshot> ReadCommand<S> for MvccByStartTs {
+    fn process_read(&mut self, snapshot: S, statistics: &mut Statistics) -> Result<ProcessResult> {
         let mut reader = MvccReader::new(
             snapshot,
             Some(ScanMode::Forward),
