+++ src/storage/txn/commands/pause.rs
@@ -2,8 +2,7 @@
 
 use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand};
-use crate::storage::txn::process::WriteResult;
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand, WriteResult};
 use crate::storage::txn::Result;
 use crate::storage::{ProcessResult, Snapshot, Statistics};
 use kvproto::kvrpcpb::ExtraOp;
