+++ src/storage/txn/commands/prewrite_pessimistic.rs
@@ -4,8 +4,7 @@ use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
 use crate::storage::mvcc::MvccTxn;
 use crate::storage::mvcc::{Error as MvccError, ErrorInner as MvccErrorInner};
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand};
-use crate::storage::txn::process::WriteResult;
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand, WriteResult};
 use crate::storage::txn::Error;
 use crate::storage::txn::Result;
 use crate::storage::types::PrewriteResult;
