+++ src/storage/txn/commands/resolve_lock_lite.rs
@@ -3,8 +3,9 @@
 use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
 use crate::storage::mvcc::MvccTxn;
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand};
-use crate::storage::txn::process::{ReleasedLocks, WriteResult};
+use crate::storage::txn::commands::{
+    Command, CommandExt, ReleasedLocks, TypedCommand, WriteCommand, WriteResult,
+};
 use crate::storage::txn::Result;
 use crate::storage::{ProcessResult, Snapshot, Statistics};
 use kvproto::kvrpcpb::ExtraOp;
