+++ src/storage/txn/commands/scan_lock.rs
@@ -5,7 +5,7 @@ use crate::storage::txn::commands::{Command, CommandExt, ReadCommand, TypedComma
 use crate::storage::txn::sched_pool::tls_collect_keyread_histogram_vec;
 use crate::storage::txn::LockInfo;
 use crate::storage::txn::{ProcessResult, Result};
-use crate::storage::{Engine, ScanMode, Statistics};
+use crate::storage::{ScanMode, Snapshot, Statistics};
 use txn_types::{Key, TimeStamp};
 
 command! {
@@ -37,12 +37,8 @@ impl CommandExt for ScanLock {
     gen_lock!(empty);
 }
 
-impl<E: Engine> ReadCommand<E> for ScanLock {
-    fn process_read(
-        &mut self,
-        snapshot: E::Snap,
-        statistics: &mut Statistics,
-    ) -> Result<ProcessResult> {
+impl<S: Snapshot> ReadCommand<S> for ScanLock {
+    fn process_read(&mut self, snapshot: S, statistics: &mut Statistics) -> Result<ProcessResult> {
         let mut reader = MvccReader::new(
             snapshot,
             Some(ScanMode::Forward),
