+++ src/storage/txn/commands/txn_heart_beat.rs
@@ -3,8 +3,7 @@
 use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
 use crate::storage::mvcc::MvccTxn;
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand};
-use crate::storage::txn::process::WriteResult;
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand, WriteCommand, WriteResult};
 use crate::storage::txn::Result;
 use crate::storage::{ProcessResult, Snapshot, Statistics, TxnStatus};
 use kvproto::kvrpcpb::ExtraOp;
